# docs/openapi.yaml
openapi: 3.0.3
info:
  title: A2A API Gateway
  description: |
    نظام تكامل شامل للتطبيقات (Application-to-Application) يربط بين التطبيقات المحلية ونظام الأوتوبيلوت وبوت تيليجرام.
    
    ## الميزات الرئيسية
    - API Gateway موحد لإدارة جميع الاتصالات
    - Message Broker للتواصل غير المتزامن
    - تكامل الأوتوبيلوت لإدارة المهام التلقائية
    - تكامل تيليجرام للتفاعل مع المستخدمين
    - نظام مراقبة شامل مع Prometheus و Grafana
    - أمان متقدم مع JWT و Rate Limiting
  version: 1.0.0
  contact:
    name: AuraOS Team
    email: support@auraos.com
    url: https://auraos.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.auraos.com
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/webhooks/telegram:
    post:
      summary: Telegram webhook
      description: Receive Telegram bot updates
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                update_id:
                  type: integer
                  example: 123456
                message:
                  type: object
                  properties:
                    message_id:
                      type: integer
                      example: 1
                    from:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 123456789
                        username:
                          type: string
                          example: john_doe
                        first_name:
                          type: string
                          example: John
                    chat:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 123456789
                        type:
                          type: string
                          example: private
                    text:
                      type: string
                      example: Hello Bot!
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messageId:
                    type: string
                    example: msg_1234567890_abc123
                  timestamp:
                    type: string
                    format: date-time
  /api/webhooks/autopilot:
    post:
      summary: Autopilot webhook
      description: Receive Autopilot system events
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  example: task_completed
                taskId:
                  type: string
                  example: task_123456
                status:
                  type: string
                  example: completed
                result:
                  type: object
                timestamp:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messageId:
                    type: string
                    example: msg_1234567890_abc123
                  timestamp:
                    type: string
                    format: date-time
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key for service authentication
...